+++ inst/examples/ex-diceleraas.R
@@ -1,0 +1,6 @@
+## Ceramic data
+## (data from Desachy 2004)
+data("compiegne", package = "folio")
+
+## Plot a Dice-Leraas diagram
+plot_diceleraas(compiegne)